+++ includes/avatar-privacy/avatar-handlers/default-icons/class-identicon-icon-provider.php
@@ -24,26 +24,28 @@
  * @license http://www.gnu.org/licenses/gpl-2.0.html
  */
 
-namespace Avatar_Privacy\Default_Icons;
+namespace Avatar_Privacy\Avatar_Handlers\Default_Icons;
 
 use Avatar_Privacy\Data_Storage\Filesystem_Cache;
+
 /**
- * An icon provider for "retro" 8-bit style icons.
+ * An icon provider for "aleavatar" icons.
  *
  * @since 1.0.0
+ * @since 2.0.0 Moved to Avatar_Privacy\Avatar_Handlers\Default_Icons
  *
  * @author Peter Putzer <github@mundschenk.at>
  */
-class Retro_Icon_Provider extends Generating_Icon_Provider {
+class Identicon_Icon_Provider extends Generating_Icon_Provider {
 
 	/**
 	 * Creates a new instance.
 	 *
-	 * @param Generator\Retro  $generator   A generator instance.
-	 * @param Filesystem_Cache $file_cache  The file cache handler.
+	 * @param Generators\Jdenticon $generator   A generator instance.
+	 * @param Filesystem_Cache     $file_cache  The file cache handler.
 	 */
-	public function __construct( Generator\Retro $generator, Filesystem_Cache $file_cache ) {
-		parent::__construct( $generator, $file_cache, [ 'retro' ] );
+	public function __construct( Generators\Jdenticon $generator, Filesystem_Cache $file_cache ) {
+		parent::__construct( $generator, $file_cache, [ 'identicon' ] );
 	}
 
 	/**
@@ -55,6 +57,6 @@ class Retro_Icon_Provider extends Generating_Icon_Provider {
 	 * @return string
 	 */
 	protected function get_filename( $identity, $size ) {
-		return "retro/{$this->get_sub_dir( $identity )}/{$identity}.svg";
+		return "identicon/{$this->get_sub_dir( $identity )}/{$identity}.svg";
 	}
 }
