+++ includes/avatar-privacy/avatar-handlers/default-icons/class-retro-icon-provider.php
@@ -24,27 +24,27 @@
  * @license http://www.gnu.org/licenses/gpl-2.0.html
  */
 
-namespace Avatar_Privacy\Default_Icons;
+namespace Avatar_Privacy\Avatar_Handlers\Default_Icons;
 
 use Avatar_Privacy\Data_Storage\Filesystem_Cache;
-
 /**
- * An icon provider for "aleavatar" icons.
+ * An icon provider for "retro" 8-bit style icons.
  *
  * @since 1.0.0
+ * @since 2.0.0 Moved to Avatar_Privacy\Avatar_Handlers\Default_Icons
  *
  * @author Peter Putzer <github@mundschenk.at>
  */
-class Identicon_Icon_Provider extends Generating_Icon_Provider {
+class Retro_Icon_Provider extends Generating_Icon_Provider {
 
 	/**
 	 * Creates a new instance.
 	 *
-	 * @param Generator\Jdenticon $generator   A generator instance.
-	 * @param Filesystem_Cache    $file_cache  The file cache handler.
+	 * @param Generators\Retro $generator   A generator instance.
+	 * @param Filesystem_Cache $file_cache  The file cache handler.
 	 */
-	public function __construct( Generator\Jdenticon $generator, Filesystem_Cache $file_cache ) {
-		parent::__construct( $generator, $file_cache, [ 'identicon' ] );
+	public function __construct( Generators\Retro $generator, Filesystem_Cache $file_cache ) {
+		parent::__construct( $generator, $file_cache, [ 'retro' ] );
 	}
 
 	/**
@@ -56,6 +56,6 @@ class Identicon_Icon_Provider extends Generating_Icon_Provider {
 	 * @return string
 	 */
 	protected function get_filename( $identity, $size ) {
-		return "identicon/{$this->get_sub_dir( $identity )}/{$identity}.svg";
+		return "retro/{$this->get_sub_dir( $identity )}/{$identity}.svg";
 	}
 }
