+++ includes/avatar-privacy/avatar-handlers/default-icons/generators/class-wavatar.php
@@ -25,7 +25,7 @@
  * @license http://www.gnu.org/licenses/gpl-2.0.html
  */
 
-namespace Avatar_Privacy\Default_Icons\Generator;
+namespace Avatar_Privacy\Avatar_Handlers\Default_Icons\Generators;
 
 use Avatar_Privacy\Tools\Images as Image_Tools;
 
@@ -33,6 +33,7 @@ use Avatar_Privacy\Tools\Images as Image_Tools;
  * A wavatar generator.
  *
  * @since 1.0.0
+ * @since 2.0.0 Moved to Avatar_Privacy\Avatar_Handlers\Default_Icons\Generators
  */
 class Wavatar extends PNG_Generator {
